@@ -7,8 +7,6 @@
     "start": "node dist/app.js",
     "postinstall": "tsc",
     "dev": "nodemon src/app.ts"
-    
-     
   },
   "author": "shamim reza",
   "license": "ISC",
@@ -24,14 +22,14 @@
   "dependencies": {
     "bcrypt": "^5.0.1",
     "body-parser": "^1.19.0",
-    "connect-mongodb-session": "^3.1.1",
+    "connect-mongodb-session": "^3.0.0",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "express-session": "^1.17.2",
     "express-validator": "^6.12.0",
     "jsonwebtoken": "^8.5.1",
-    "mongoose": "^6.3.5",
+    "mongoose": "^5.12.5",
     "rand-token": "^1.0.1",
     "redis": "^3.1.2",
     "ts-node": "^10.2.1",
@@ -39,6 +37,6 @@
   },
   "engines": {
     "npm": ">=6.x",
-    "node": ">=12.x"
+    "node": ">=16.x"
   }
 }
